+++ app/src/main/java/com/insa/mygamelist/data/model/GameUpdated.kt
@@ -1,4 +1,4 @@
-package com.insa.mygamelist.data
+package com.insa.mygamelist.data.model
 
 import kotlinx.serialization.Serializable
