+++ app/src/main/java/com/insa/mygamelist/ui/components/ListOfGames.kt
@@ -1,4 +1,4 @@
-package com.insa.mygamelist.ui.views
+package com.insa.mygamelist.ui.components
 
 import android.util.Log
 import androidx.compose.foundation.clickable
@@ -12,9 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.insa.mygamelist.data.GameUpdated
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
+import com.insa.mygamelist.data.model.GameUpdated
 
 
 @Composable
